Applied changes to AddWorkoutToLibraryDialog such ass keyboard options, color changes to cancel button, and exit animations to fabs when user clicks on Add Workout to library.
